Remove need to load django settings
By calling settings.configure if the Django settings have not already
been configured in the makers.py code, remove the need to load django
settings before importing dmdj in an arbitrary python context. This
removes the need to define a secret key in the settings.py file and the
need to set the DJANGO_SETTINGS_MODULE in the test code.

Additionally, the module generation code is updated to add a final
`models` module to the end of the module path, in order to match what
would normally be expected in a Django app and the pattern from
data-models-sqlalchemy.

The README file (which was from the original pedsnetcdms repo) is
massively updated (and condensed) so that it now represents information
about this actual package, instead of the old one.

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>